\name{bigBits-package}
\alias{bigBits-package}
\alias{bigBits}
\docType{package}
\title{
\packageTitle{bigBits}
}
\description{
\packageDescription{bigBits}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bigBits}
\packageIndices{bigBits}

}
\author{
\packageAuthor{bigBits}

Maintainer: \packageMaintainer{bigBits}
}
\references{
\url{https://en.wikipedia.org/wiki/Two's_complement} 
}
\keyword{ package }
\seealso{

\code{\link[base]{bitwAnd}} and other "bitw*" functions

}

